requirements.txt
langchain>=1.2.10
langchain-openai>=1.1.10
langgraph>=1.0.8
requests>=2.32.5
python-dotenv>=1.2.1
pydantic>=2.12.5
pytz>=2025.2
yt-dlp>=2024.10.7
youtube-transcript-api>=0.6.2
